# services/api/openapi/cart/v1/cart.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "cart/v1/cart.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CartService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/carts/{sessionId}": {
      "get": {
        "operationId": "CartService_GetCart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "CartService"
        ]
      }
    },
    "/v1/carts/{sessionId}/items": {
      "post": {
        "operationId": "CartService_AddItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AddItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CartServiceAddItemBody"
            }
          }
        ],
        "tags": [
          "CartService"
        ]
      }
    },
    "/v1/carts/{sessionId}/items/{productUuid}": {
      "delete": {
        "operationId": "CartService_RemoveItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RemoveItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "productUuid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "CartService"
        ]
      },
      "patch": {
        "operationId": "CartService_UpdateItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "productUuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CartServiceUpdateItemBody"
            }
          }
        ],
        "tags": [
          "CartService"
        ]
      }
    }
  },
  "definitions": {
    "CartServiceAddItemBody": {
      "type": "object",
      "properties": {
        "productUuid": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "CartServiceUpdateItemBody": {
      "type": "object",
      "properties": {
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AddItemResponse": {
      "type": "object",
      "properties": {
        "cart": {
          "$ref": "#/definitions/v1Cart"
        }
      }
    },
    "v1Cart": {
      "type": "object",
      "properties": {
        "sessionId": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1CartItem"
          }
        },
        "total": {
          "type": "string"
        }
      }
    },
    "v1CartItem": {
      "type": "object",
      "properties": {
        "productUuid": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "unitPrice": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "lineTotal": {
          "type": "string"
        }
      }
    },
    "v1GetCartResponse": {
      "type": "object",
      "properties": {
        "cart": {
          "$ref": "#/definitions/v1Cart"
        }
      }
    },
    "v1RemoveItemResponse": {
      "type": "object",
      "properties": {
        "cart": {
          "$ref": "#/definitions/v1Cart"
        }
      }
    },
    "v1UpdateItemResponse": {
      "type": "object",
      "properties": {
        "cart": {
          "$ref": "#/definitions/v1Cart"
        }
      }
    }
  }
}